+++ app/src/main/kotlin/com/lardis/i_larin/module/storage/entities/MessageModel.kt
@@ -2,6 +2,6 @@ package com.lardis.i_larin.module.storage.entities
 
 data class MessageModel constructor(val id: Long?,
                                     val message: String,
-                                    val idAuthor: Long,
+                                    val idAuthor: Int,
                                     val idDialog: Long,
                                     val timeCreate: Long)
